Fixes fields_for with deeply nested has_many associations
This code is not 100% perfect. If you would have an attribute name build up with exactly 24 characters containing only [0-9a-f] the code would not recognize it as an attribute name but would think it is an id for the nested attributes array. The same goes for when you would have attributes named new_\d.